# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Finance Tracker",
    "description": "Track income, expenses, budgets, and recurring transactions.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/dashboard": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Dashboard",
        "description": "Return financial KPIs and recent activity.",
        "operationId": "get_dashboard_dashboard_get",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/csv/import": {
      "post": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Import Transactions Csv",
        "operationId": "import_transactions_csv_csv_import_post",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_transactions_csv_csv_import_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/budget/": {
      "post": {
        "tags": [
          "Budget"
        ],
        "summary": "Set Budget",
        "description": "Create or update a monthly budget for a category.",
        "operationId": "set_budget_budget__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/budget/status": {
      "get": {
        "tags": [
          "Budget"
        ],
        "summary": "Budget Status",
        "description": "Return spending vs budget for each configured category.",
        "operationId": "budget_status_budget_status_get",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BudgetStatus"
                  },
                  "title": "Response Budget Status Budget Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recurring/": {
      "get": {
        "tags": [
          "Recurring"
        ],
        "summary": "List Recurring",
        "description": "List all recurring transaction templates.",
        "operationId": "list_recurring_recurring__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecurringOut"
                  },
                  "type": "array",
                  "title": "Response List Recurring Recurring  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Recurring"
        ],
        "summary": "Create Recurring",
        "description": "Create a recurring transaction template.",
        "operationId": "create_recurring_recurring__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/recurring/{rec_id}": {
      "delete": {
        "tags": [
          "Recurring"
        ],
        "summary": "Delete Recurring",
        "description": "Cancel a recurring transaction template.",
        "operationId": "delete_recurring_recurring__rec_id__delete",
        "parameters": [
          {
            "name": "rec_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rec Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/savings/": {
      "get": {
        "tags": [
          "Savings"
        ],
        "summary": "List Saving Goals",
        "operationId": "list_saving_goals_savings__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavingGoalOut"
                  },
                  "type": "array",
                  "title": "Response List Saving Goals Savings  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Savings"
        ],
        "summary": "Create Saving Goal",
        "operationId": "create_saving_goal_savings__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingGoalCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingGoalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/savings/{goal_id}/add": {
      "patch": {
        "tags": [
          "Savings"
        ],
        "summary": "Add To Saving Goal",
        "operationId": "add_to_saving_goal_savings__goal_id__add_patch",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingGoalAdd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingGoalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/savings/deposit": {
      "post": {
        "tags": [
          "Savings"
        ],
        "summary": "Deposit To Savings",
        "operationId": "deposit_to_savings_savings_deposit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingsTransactionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingsTransactionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/savings/withdraw": {
      "post": {
        "tags": [
          "Savings"
        ],
        "summary": "Withdraw From Savings",
        "operationId": "withdraw_from_savings_savings_withdraw_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavingsTransactionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavingsTransactionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_transactions_csv_csv_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_transactions_csv_csv_import_post"
      },
...
      "BudgetCreate": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "monthly_limit": {
            "type": "number",
            "title": "Monthly Limit"
          }
        },
        "type": "object",
        "required": [
          "category",
          "monthly_limit"
        ],
        "title": "BudgetCreate"
      },
      "BudgetOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "monthly_limit": {
            "type": "number",
            "title": "Monthly Limit"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "monthly_limit"
        ],
        "title": "BudgetOut"
      },
      "BudgetStatus": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "monthly_limit": {
            "type": "number",
            "title": "Monthly Limit"
          },
          "current_spending": {
            "type": "number",
            "title": "Current Spending"
          },
          "projected_spending": {
            "type": "number",
            "title": "Projected Spending"
          },
          "days_left": {
            "type": "integer",
            "title": "Days Left"
          },
          "exceeded": {
            "type": "boolean",
            "title": "Exceeded"
          }
        },
        "type": "object",
        "required": [
          "category",
          "monthly_limit",
          "current_spending",
          "projected_spending",
          "days_left",
          "exceeded"
        ],
        "title": "BudgetStatus"
      },
...
      "DashboardOut": {
        "properties": {
          "total_income": {
            "type": "number",
            "title": "Total Income"
          },
          "total_expenses": {
            "type": "number",
            "title": "Total Expenses"
          },
          "balance": {
            "type": "number",
            "title": "Balance"
          },
          "total_savings": {
            "type": "number",
            "title": "Total Savings"
          },
          "monthly_income": {
            "type": "number",
            "title": "Monthly Income"
          },
          "monthly_expenses": {
            "type": "number",
            "title": "Monthly Expenses"
          },
          "carried_savings": {
            "type": "number",
            "title": "Carried Savings"
          },
          "total_transactions": {
            "type": "integer",
            "title": "Total Transactions"
          },
          "top_spending_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Spending Category"
          },
          "recent_activity": {
            "items": {
              "$ref": "#/components/schemas/TransactionOut"
            },
            "type": "array",
            "title": "Recent Activity"
          }
        },
        "type": "object",
        "required": [
          "total_income",
          "total_expenses",
          "balance",
          "total_savings",
          "monthly_income",
          "monthly_expenses",
          "carried_savings",
          "total_transactions",
          "top_spending_category",
          "recent_activity"
        ],
        "title": "DashboardOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecurringCreate": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "transaction_type": {
            "type": "string",
            "title": "Transaction Type"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          },
          "next_due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Due Date"
          }
        },
        "type": "object",
        "required": [
          "description",
          "amount",
          "transaction_type",
          "frequency"
        ],
        "title": "RecurringCreate"
      },
      "RecurringOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "transaction_type": {
            "type": "string",
            "title": "Transaction Type"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          },
          "next_due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Due Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "description",
          "amount",
          "transaction_type",
          "category",
          "frequency"
        ],
        "title": "RecurringOut"
      },
      "SavingGoalAdd": {
        "properties": {
          "amount": {
            "type": "number",
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "SavingGoalAdd"
      },
      "SavingGoalCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_amount": {
            "type": "number",
            "title": "Target Amount"
          }
        },
        "type": "object",
        "required": [
          "name",
          "target_amount"
        ],
        "title": "SavingGoalCreate"
      },
      "SavingGoalOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_amount": {
            "type": "number",
            "title": "Target Amount"
          },
          "current_amount": {
            "type": "number",
            "title": "Current Amount"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "target_amount",
          "current_amount"
        ],
        "title": "SavingGoalOut"
      },
...
      "SavingsTransactionCreate": {
        "properties": {
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "goal_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal Id"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "SavingsTransactionCreate"
      },
      "SavingsTransactionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "goal_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "amount",
          "reason",
          "goal_id",
          "timestamp"
        ],
        "title": "SavingsTransactionOut"
      },
...
      "TransactionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "transaction_type": {
            "type": "string",
            "title": "Transaction Type"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "auto_tagged": {
            "type": "boolean",
            "title": "Auto Tagged"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "id",
          "amount",
          "description",
          "transaction_type",
          "category",
          "auto_tagged",
          "timestamp"
        ],
        "title": "TransactionOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}